requirements.txt
streamlit
pandas
streamlit-aggrid
numpy
google-cloud-aiplatform
openpyxl
